Moved 'autocomplete_on_id' from search to xref crud and router
This commit moves the autocomplete functionality from 'search_router.py' and 'search_crud.py' to 'cross_reference_router.py' and 'cross_reference_crud.py'
