+++ agr_literature_service/api/crud/cross_reference_crud.py
@@ -178,3 +178,22 @@ def show_changesets(db: Session, cross_reference_id: int):
             }
         )
     return history
+
+
+def autocomplete_on_id(prefix: str, query: str, return_prefix: bool, db: Session):
+    string_before_id = ""
+    if prefix == "WB":
+        string_before_id = "WBPaper"
+    if query.startswith(string_before_id):
+        query = query[len(string_before_id)]
+    matching_xrefs_query = db.query(CrossReferenceModel.curie).filter(
+        CrossReferenceModel.curie.like(f"{prefix}:{string_before_id}{query}%")
+    )
+    matching_xrefs_count = matching_xrefs_query.count()
+    matching_xrefs = matching_xrefs_query.limit(20).all()
+    matching_curies = ["".join(matching_xref.curie.split(":")[1:]) if not return_prefix else matching_xref.curie for
+                       matching_xref in matching_xrefs]
+    if matching_xrefs_count > 20:
+        matching_curies.append("more ...")
+    matching_curies_plain_text = "\n".join(matching_curies)
+    return matching_curies_plain_text
